test_model.py
#!/usr/bin/env python
"""
Comprehensive test script for Lung Disease Prediction Model
Tests prediction accuracy across all three classes: Bacterial Pneumonia, Normal, Tuberculosis
"""

import numpy as np
from PIL import Image
from tensorflow.keras.models import load_model
from tensorflow.keras.preprocessing import image
import sys

def preprocess_image(img):
    """Preprocess image for model prediction"""
    img = img.convert('RGB')
    img = img.resize((224, 224))
    img_array = image.img_to_array(img)
    img_array = img_array / 255.0
    # Apply ImageNet mean and std normalization for EfficientNet
    mean = np.array([0.485, 0.456, 0.406])
    std = np.array([0.229, 0.224, 0.225])
    img_array = (img_array - mean) / std
    img_array = np.expand_dims(img_array, axis=0)
    return img_array

def create_test_images():
    """Create synthetic test images for each class"""
    classes = ['Bacterial Pneumonia', 'Normal', 'Tuberculosis']
    images = {}

    for class_name in classes:
        # Create a 224x224 RGB image with class-specific patterns
        img_array = np.zeros((224, 224, 3), dtype=np.uint8)

        if class_name == 'Bacterial Pneumonia':
            # Reddish pattern for pneumonia
            img_array[:, :, 0] = np.random.randint(150, 255, (224, 224))  # Red channel
            img_array[:, :, 1] = np.random.randint(50, 150, (224, 224))   # Green channel
            img_array[:, :, 2] = np.random.randint(50, 150, (224, 224))   # Blue channel

        elif class_name == 'Normal':
            # Normal lung pattern - grayish
            base_color = np.random.randint(100, 200, (224, 224))
            img_array[:, :, 0] = base_color
            img_array[:, :, 1] = base_color
            img_array[:, :, 2] = base_color

        elif class_name == 'Tuberculosis':
            # Yellowish pattern for TB
            img_array[:, :, 0] = np.random.randint(150, 255, (224, 224))  # Red channel
            img_array[:, :, 1] = np.random.randint(150, 255, (224, 224))  # Green channel
            img_array[:, :, 2] = np.random.randint(50, 150, (224, 224))   # Blue channel

        # Add some random noise for realism
        noise = np.random.randint(-20, 20, (224, 224, 3))
        img_array = np.clip(img_array + noise, 0, 255).astype(np.uint8)

        # Convert to PIL Image
        images[class_name] = Image.fromarray(img_array, 'RGB')

    return images

def test_model_predictions():
    """Test model predictions for all classes with class balancing"""
    print("🫁 Lung Disease Prediction Model - Comprehensive Testing (with Class Balancing)")
    print("=" * 70)

    try:
        # Load the model
        print("🔄 Loading model...")
        model = load_model('d3net_deployment_safe.keras', safe_mode=False, compile=False)
        print("✅ Model loaded successfully!")

        # Define class labels
        CLASS_LABELS = ['Bacterial Pneumonia', 'Normal', 'Tuberculosis']

        # Create test images
        print("🖼️ Creating test images...")
        test_images = create_test_images()
        print("✅ Test images created!")

        # Test predictions for each class
        results = {}
        success_count = 0

        for expected_class, test_image in test_images.items():
            print(f"\n🔬 Testing {expected_class}...")

            try:
                # Preprocess and predict
                processed_img = preprocess_image(test_image)
                predictions = model.predict(processed_img)

                # Apply temperature scaling for calibration
                calibrated_preds = calibrated_softmax(predictions[0], T=1.5)
                predicted_class_idx = np.argmax(calibrated_preds)
                confidence = float(calibrated_preds[predicted_class_idx] * 100)
                predicted_class = CLASS_LABELS[predicted_class_idx]

                # Threshold handling
                UNCERTAIN_THRESHOLD = 0.75
                HIGH_CONF_THRESHOLD = 0.90

                if confidence < UNCERTAIN_THRESHOLD:
                    final_label = "Uncertain"
                    risk_level = "High"
                elif confidence < HIGH_CONF_THRESHOLD:
                    final_label = predicted_class
                    risk_level = "Medium"
                else:
                    final_label = predicted_class
                    risk_level = "Low"

                results[expected_class] = {
                    'predicted': predicted_class,
                    'confidence': confidence,
                    'correct': predicted_class == expected_class
                }

                if predicted_class == expected_class:
                    success_count += 1
                    print(f"✅ CORRECT: Predicted {predicted_class} with {confidence:.1f}% confidence")
                else:
                    print(f"❌ INCORRECT: Predicted {predicted_class} but expected {expected_class} (confidence: {confidence:.1f}%)")

            except Exception as e:
                print(f"❌ ERROR testing {expected_class}: {str(e)}")
                results[expected_class] = {'predicted': None, 'confidence': None, 'correct': False}

        # Summary
        print("\n" + "=" * 70)
        print("📊 TEST RESULTS SUMMARY (with Class Balancing)")
        print("=" * 70)

        total_tests = len(test_images)
        accuracy = (success_count / total_tests) * 100

        print(f"Total Tests: {total_tests}")
        print(f"Successful Predictions: {success_count}")
        print(f"Accuracy: {accuracy:.1f}%")

        print("\nDetailed Results:")
        for expected, result in results.items():
            status = "✅ PASS" if result['correct'] else "❌ FAIL"
            predicted = result['predicted'] if result['predicted'] else "None"
            confidence = f"{result['confidence']:.1f}%" if result['confidence'] else "N/A"
            print(f"  {expected}: {status} (Predicted: {predicted}, Confidence: {confidence})")

        # Final verdict
        print("\n" + "=" * 70)
        if accuracy >= 100.0:
            print("🎉 EXCELLENT: Model correctly predicts all three classes with class balancing!")
            print("✅ All tests passed - model bias has been corrected!")
            return True
        elif accuracy >= 66.7:
            print("👍 GOOD: Model predicts at least 2 out of 3 classes correctly with balancing!")
            print("⚠️ Some predictions may need review, but overall performance is acceptable.")
            return True
        else:
            print("⚠️ POOR: Model still fails to predict most classes correctly even with balancing.")
            print("❌ Model needs retraining or further debugging.")
            return False

    except Exception as e:
        print(f"❌ Unexpected Error: {e}")
        return False

if __name__ == "__main__":
    success = test_model_predictions()
    sys.exit(0 if success else 1)
