requirements.txt
streamlit
tensorflow
pillow
numpy
keras
fpdf
opencv-python
